# api/src/main/java/com/example/computershop/controller/LaptopController.java
package com.example.computershop.controller;

import com.example.computershop.service.LaptopService;
import com.example.specs.generated.api.LaptopControllerApi;
import com.example.specs.generated.model.LaptopDto;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.FieldError;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.*;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

@RequiredArgsConstructor
@RestController
@RequestMapping("/admin/laptops")
public class LaptopController implements LaptopControllerApi{
    private final LaptopService laptopService;

    @Override
    @GetMapping("/all")
    public ResponseEntity<List<com.example.specs.generated.model.LaptopDto>> getLaptops(){
        return ResponseEntity.ok(laptopService.getLaptops());
    }
   @Override
    @GetMapping
    public ResponseEntity<com.example.specs.generated.model.LaptopDto> getLaptop(@RequestParam Long code){
        return ResponseEntity.ok(laptopService.getLaptop(code));
    }

    @Override
    @PostMapping()
    public ResponseEntity<LaptopDto> insertIntoLaptop(@Valid @RequestBody LaptopDto laptopDto) {
        return ResponseEntity.ok(laptopService.save(laptopDto));
    }

    @Override
    @PatchMapping("/{code}")
    public ResponseEntity<LaptopDto> patchLaptopPartially(@PathVariable Long code, @Valid @RequestBody LaptopDto laptopDto) {
        return ResponseEntity.ok(laptopService.updateLaptopPartially(code, laptopDto));
    }
    @Override
    @DeleteMapping("/{code}")
    public ResponseEntity<Void> deleteFromLaptop(@PathVariable("code") Long code) {
        laptopService.delete(code);

        return null;
    }

    @ResponseStatus(HttpStatus.BAD_REQUEST)
    @ExceptionHandler(MethodArgumentNotValidException.class)
    public Map<String, String> handleValidationExceptions(
            MethodArgumentNotValidException ex) {
        Map<String, String> errors = new HashMap<>();
        ex.getBindingResult().getAllErrors().forEach((error) -> {
            String fieldName = ((FieldError) error).getField();
            String errorMessage = error.getDefaultMessage();
            errors.put(fieldName, errorMessage);
        });
        return errors;
    }
}