# specs/src/main/resources/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/admin/printers": {
      "get": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "getPrinter",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "insertIntoPrinter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Printer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      }
    },
    "/admin/pcs": {
      "get": {
        "tags": [
          "pc-controller"
        ],
        "operationId": "getPc",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pc"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pc-controller"
        ],
        "operationId": "insertIntoPc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pc"
                }
              }
            }
          }
        }
      }
    },
    "/admin/laptops": {
      "get": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "getLaptop",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "insertIntoLaptop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Laptop"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/printers/{code}": {
      "delete": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "deleteFromPrinter",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "patchPrinterPartially",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Printer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      }
    },
    "/admin/pcs/{code}": {
      "delete": {
        "tags": [
          "pc-controller"
        ],
        "operationId": "deleteFromPc",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "pc-controller"
        ],
        "operationId": "patchPcPartially",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pc"
                }
              }
            }
          }
        }
      }
    },
    "/admin/laptops/{code}": {
      "delete": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "deleteFromLaptop",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "patchLaptopPartially",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Laptop"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/cart/user/{username}": {
      "get": {
        "tags": [
          "cart-consumer-controller"
        ],
        "operationId": "getCustomerCartForCustomer",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cart"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cart-consumer-controller"
        ],
        "operationId": "clearCart",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/printers/all": {
      "get": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "getPrinters",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Printer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/laptops/all": {
      "get": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "getLaptops",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Laptop"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/cart/{orderId}": {
      "delete": {
        "tags": [
          "cart-consumer-controller"
        ],
        "operationId": "deleteCustomerOneFromCart",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cart": {
        "required": [
          "model",
          "price",
          "username"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Printer": {
        "required": [
          "color",
          "price",
          "type"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Pc": {
        "required": [
          "cd",
          "hd",
          "price",
          "ram",
          "speed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "hd": {
            "type": "number",
            "format": "double"
          },
          "cd": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Laptop": {
        "required": [
          "hd",
          "model",
          "price",
          "ram",
          "screen",
          "speed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "hd": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "screen": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}